Separated page UI from table container
