+++ src/js/containers/router/RouterRoutes.js
@@ -31,7 +31,7 @@ const AgencyProfileV2 = React.lazy(() => import('containers/agency/v2/AgencyCont
 const Covid19Container = React.lazy(() => import('containers/covid19/Covid19Container').then((comp) => comp));
 const DataSourcesAndMethodologiesPage = React.lazy(() => import('components/covid19/DataSourcesAndMethodologiesPage').then((comp) => comp));
 const AboutTheDataPage = React.lazy(() => import('components/aboutTheData/AboutTheDataPage').then((comp) => comp));
-const AgenciesDetailContainer = React.lazy(() => import('containers/aboutTheData/AgenciesDetailContainer').then((comp) => comp));
+const AgencyDetailsPage = React.lazy(() => import('components/aboutTheData/AgencyDetailsPage').then((comp) => comp));
 const ErrorPage = React.lazy(() => import('components/errorPage/ErrorPage').then((comp) => comp));
 
 // /* eslint-disable import/prefer-default-export */
@@ -175,8 +175,8 @@ export const routes = [
         hide: !kGlobalConstants.DEV && !kGlobalConstants.QAT // Not DEV and not QAT === Production, so we hide
     },
     {
-        path: '/about-the-data/agency/:agencyId',
-        component: AgenciesDetailContainer,
+        path: '/about-the-data/agency/:agencyCode',
+        component: AgencyDetailsPage,
         exact: true,
         hide: !kGlobalConstants.DEV && !kGlobalConstants.QAT // Not DEV and not QAT === Production, so we hide
     },
