+++ src/js/helpers/agencyV2Helper.js
@@ -1,5 +1,5 @@
 /**
- * agencyHelper.js
+ * agencyV2Helper.js
  * Created by Lizzie Salita 5/26/20
  */
 
@@ -19,4 +19,8 @@ export const fetchSpendingByCategory = (agencyId, type, params) => apiRequest({
 
 export const fetchBudgetaryResources = (agencyId) => apiRequest({
     url: `v2/agency/${agencyId}/budgetary_resources/`
+});
+
+export const fetchAgencyOverview = (code, fy) => apiRequest({
+    url: `v2/agency/${code}/${fy || ''}`
 });
